+++ js/utils.js
@@ -1,6 +1,4 @@
 import {main} from "./main";
-import {PIXEL_HUNTER} from "./data/game";
-import {resize} from "./data/resize";
 
 export const getElementFromTemplate = (text) => {
   const element = document.createElement(`div`);
@@ -19,19 +17,4 @@ export const checkQuestion = (question) => {
     value = q.value;
     return q.type === `radio` && q.checked;
   }) ? value : false;
-};
-
-export const addImages = (images, state, frame) => {
-  [...images].forEach((it, i) => {
-    const img = new Image();
-    img.addEventListener(`load`, function () {
-      const naturalSize = {width: img.width, height: img.height};
-      const resolution = resize(frame, naturalSize);
-      img.width = resolution.width;
-      img.height = resolution.height;
-      it.appendChild(img);
-    });
-    img.src = PIXEL_HUNTER[state.level].answers[i].img;
-    img.alt = `Option ${i + 1}`;
-  });
 };
